LPS-94217 Move AssetRendererFactorys from *.asset to *.asset.model packages
