+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/asset/model/BlogsEntryAssetRenderer.java
@@ -20,7 +20,6 @@ import com.liferay.asset.kernel.model.BaseJSPAssetRenderer;
 import com.liferay.asset.util.AssetHelper;
 import com.liferay.blogs.constants.BlogsPortletKeys;
 import com.liferay.blogs.model.BlogsEntry;
-import com.liferay.blogs.web.asset.BlogsEntryAssetRendererFactory;
 import com.liferay.blogs.web.internal.security.permission.resource.BlogsEntryPermission;
 import com.liferay.blogs.web.internal.util.BlogsEntryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
