+++ modules/apps/calendar/calendar-web/src/main/java/com/liferay/calendar/web/internal/asset/model/CalendarBookingAssetRendererFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.calendar.web.internal.asset;
+package com.liferay.calendar.web.internal.asset.model;
 
 import com.liferay.asset.kernel.model.AssetRenderer;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
@@ -23,7 +23,6 @@ import com.liferay.calendar.model.Calendar;
 import com.liferay.calendar.model.CalendarBooking;
 import com.liferay.calendar.model.CalendarResource;
 import com.liferay.calendar.service.CalendarBookingLocalService;
-import com.liferay.calendar.web.internal.asset.model.CalendarBookingAssetRenderer;
 import com.liferay.calendar.web.internal.util.CalendarResourceUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
