+++ modules/apps/tasks/tasks-portlet/docroot/WEB-INF/liferay-portlet.xml
@@ -8,7 +8,7 @@
 		<social-activity-interpreter-class>com.liferay.tasks.social.TasksActivityInterpreter</social-activity-interpreter-class>
 		<user-notification-definitions>com/liferay/tasks/portlet/tasks-user-notification-definitions.xml</user-notification-definitions>
 		<user-notification-handler-class>com.liferay.tasks.notifications.TasksUserNotificationHandler</user-notification-handler-class>
-		<asset-renderer-factory>com.liferay.tasks.asset.TasksEntryAssetRendererFactory</asset-renderer-factory>
+		<asset-renderer-factory>com.liferay.tasks.asset.model.TasksEntryAssetRendererFactory</asset-renderer-factory>
 		<header-portlet-css>/tasks/css/main.css</header-portlet-css>
 		<footer-portlet-javascript>/tasks/js/main.js</footer-portlet-javascript>
 		<css-class-wrapper>tasks-portlet</css-class-wrapper>
